pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contract_risk_analyzer"
version = "0.1.0"
description = "OpenEnv environment for contract risk analysis."
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "pydantic>=2.5",
    "requests>=2.31",
    "openai>=1.30",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["server"]
py-modules = ["models", "client", "inference"]
